+++ Services/CDN/CDN.Application/Contracts/Persistence/INewsLetterRepository.cs
@@ -5,7 +5,7 @@ namespace CDN.Application.Contracts.Persistence
     public interface INewsLetterRepository
     {
         Task<NewsLetter> CreateAsync(NewsLetter newsLetter, CancellationToken cancellationToken = default);
-        NewsLetter GetWithCategory(long id);
-        Task<NewsLetter> GetAsync(long id, CancellationToken cancellationToken = default);
+        NewsLetter GetWithCategory(int id);
+        Task<NewsLetter> GetAsync(int id, CancellationToken cancellationToken = default);
     }
 }
