+++ Services/SSO/SSO.Application/Contracts/Persistence/IRoleRepository.cs
@@ -8,11 +8,11 @@ namespace SSO.Application.Contracts.Persistence
     public interface IRoleRepository
     {
         Task<List<Role>> GetAsync(CancellationToken cancellationToken);
-        Task<Role> GetAsync(long id, CancellationToken cancellationToken);
-        Task<List<Role>> GetByRoleIdsAsync(List<long> ids, CancellationToken cancellationToken);
-        Task<Role> GetWithGroupsAsync(long id, CancellationToken cancellationToken);
-        Task<Role> GetWithUsersAsync(long id, CancellationToken cancellationToken);
-        Task<bool> IsUniqueDisplayTitleAsync(long id, string displayTitle, CancellationToken cancellationToken);
+        Task<Role> GetAsync(int id, CancellationToken cancellationToken);
+        Task<List<Role>> GetByRoleIdsAsync(List<int> ids, CancellationToken cancellationToken);
+        Task<Role> GetWithGroupsAsync(int id, CancellationToken cancellationToken);
+        Task<Role> GetWithUsersAsync(int id, CancellationToken cancellationToken);
+        Task<bool> IsUniqueDisplayTitleAsync(int id, string displayTitle, CancellationToken cancellationToken);
         Task UpdateAsync(Role role, CancellationToken cancellationToken);
     }
 }
