+++ Services/SSO/SSO.Application/Contracts/Persistence/IUserRepository.cs
@@ -12,18 +12,18 @@ namespace SSO.Application.Contracts.Persistence
         Task DeleteAsync(User user, CancellationToken cancellationToken);
         Task<PaginatedResult<User>> GetAsync(string searchParam, int pageNumber, int pageSize, CancellationToken cancellationToken);
         Task<PaginatedResult<User>> GetAsync(int pageNumber, int pageSize, CancellationToken cancellationToken);
-        Task<User> GetAsync(long id, CancellationToken cancellationToken);
+        Task<User> GetAsync(int id, CancellationToken cancellationToken);
         Task UpdateAsync(User user, CancellationToken cancellationToken);
         Task<User> GetUserWithRolesAsync(string userName, string encPass, CancellationToken cancellationToken);
-        Task<User> GetUserWithRolesAsync(long id, CancellationToken cancellationToken);
-        Task<User> GetUserByPasswordAsync(long userId, string encPass, CancellationToken cancellationToken);
-        Task<User> GetWithRoleAndRefreshTokensAsync(long userId, CancellationToken cancellationToken);
-        Task<List<User>> GetByUserIdsAsync(List<long> ids, CancellationToken cancellationToken);
-        Task<User> GetUserWithGroupsAsync(long id, CancellationToken cancellationToken);
+        Task<User> GetUserWithRolesAsync(int id, CancellationToken cancellationToken);
+        Task<User> GetUserByPasswordAsync(int userId, string encPass, CancellationToken cancellationToken);
+        Task<User> GetWithRoleAndRefreshTokensAsync(int userId, CancellationToken cancellationToken);
+        Task<List<User>> GetByUserIdsAsync(List<int> ids, CancellationToken cancellationToken);
+        Task<User> GetUserWithGroupsAsync(int id, CancellationToken cancellationToken);
         Task<bool> IsUniqueUserNameAsync(string userName, CancellationToken cancellationToken);
         Task<User> GetByUserNameAsync(string userName, CancellationToken cancellationToken);
         Task<bool> IsUniqueMobileAsync(string mobile, CancellationToken cancellationToken);
-        Task<bool> IsUniqueMobileAsync(string mobile, long id, CancellationToken cancellationToken);
-        Task<PaginatedResult<User>> GetByGroupIdAsync(long groupId, string searchParam, int pageNumber, int pageSize, CancellationToken cancellationToken);
+        Task<bool> IsUniqueMobileAsync(string mobile, int id, CancellationToken cancellationToken);
+        Task<PaginatedResult<User>> GetByGroupIdAsync(int groupId, string searchParam, int pageNumber, int pageSize, CancellationToken cancellationToken);
     }
 }
