+++ Services/SSO/SSO.Application/Features/GroupFeature/Queries/GetGroupRoles/GetGroupRolesQuery.cs
@@ -4,6 +4,6 @@ namespace SSO.Application.Features.GroupFeature.Queries.GetGroupRoles
 {
     public class GetGroupRolesQuery : IRequest<GroupRolesDto>
     {
-        public long GroupId { get; set; }
+        public int GroupId { get; set; }
     }
 }
